+++ CHANGELOG.md
@@ -2,12 +2,15 @@
 
 ## Current Main
 
-### New Features
+### Other Changes
 
-- add source and description to quality dimensions ([#720])
+- road-comparison: change color of bar to grey ([#798])
 
+### Bug Fixes
 
-[#720]: https://github.com/GIScience/ohsome-quality-api/pull/720
+- road-comparison: bar for not covered roads is now shown in the plot ([#798])
+
+[#798]: https://github.com/GIScience/ohsome-quality-api/pull/798
 
 ## Release 1.3.0